requirements.txt
langgraph>=1.0.0
langchain-groq==1.0.0
python-dotenv==1.2.1